third_party/blink/web_tests/http/tests/websocket/close-code-and-reason-expected.txt

Test CloseEvent code and reason property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

ws.onclose() was called.
PASS closeEvent.wasClean is false
PASS closeEvent.code is codeAbnormalClosure
PASS closeEvent.reason is emptyString
WebSocketTest.onopen() was called with testId = 0.
WebSocketTest.onclose() was called with testId = 0.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 1005
PASS closeEvent.reason is ''
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 1005
PASS closeEvent.reason is ''
WebSocketTest.onopen() was called with testId = 1.
WebSocketTest.onclose() was called with testId = 1.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
WebSocketTest.onopen() was called with testId = 2.
WebSocketTest.onclose() was called with testId = 2.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 1000
PASS closeEvent.reason is 'ok'
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 1000
PASS closeEvent.reason is 'ok'
WebSocketTest.onopen() was called with testId = 3.
WebSocketTest.onclose() was called with testId = 3.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
WebSocketTest.onopen() was called with testId = 4.
WebSocketTest.onclose() was called with testId = 4.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
WebSocketTest.onopen() was called with testId = 5.
WebSocketTest.onclose() was called with testId = 5.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS closeEvent.reason is ''
WebSocketTest.onopen() was called with testId = 6.
WebSocketTest.onclose() was called with testId = 6.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 0
PASS closeEvent.reason is 'good bye'
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 0
PASS closeEvent.reason is 'good bye'
WebSocketTest.onopen() was called with testId = 7.
WebSocketTest.onclose() was called with testId = 7.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 10
PASS closeEvent.reason is 'good bye'
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 10
PASS closeEvent.reason is 'good bye'
WebSocketTest.onopen() was called with testId = 8.
WebSocketTest.onclose() was called with testId = 8.
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 65535
PASS closeEvent.reason is 'good bye'
PASS typeof closeEvent.wasClean is 'boolean'
PASS closeEvent.wasClean is true
PASS closeEvent.code is 65535
PASS closeEvent.reason is 'good bye'
PASS successfullyParsed is true

TEST COMPLETE
